@@ -8,7 +8,7 @@ class HintText extends StatelessWidget {
   final int maxLine;
   final double fontSize;
 
-  HintText({this.label, this.maxLine = 1, this.fontSize = fontXSmall});
+  HintText({required this.label, this.maxLine = 1, this.fontSize = fontXSmall});
 
   @override
   Widget build(BuildContext context) {
@@ -17,7 +17,13 @@ class HintText extends StatelessWidget {
         label,
         maxLines: maxLine,
         overflow: TextOverflow.ellipsis,
-        style: TextStyle(color: hintTextColor, fontFamily: 'AvertaSemibold', fontSize: fontSize, fontWeight: fontWeightSemiBold, letterSpacing: 0.28, height: 1.5),
+        style: TextStyle(
+            color: hintTextColor,
+            fontFamily: 'AvertaSemibold',
+            fontSize: fontSize,
+            fontWeight: fontWeightSemiBold,
+            letterSpacing: 0.28,
+            height: 1.5),
       ),
     );
   }
